Tìm kiếm bài đăng theo yêu cầu sử dựng sequelize operator #51

+++ server/src/models/post.js
@@ -41,8 +41,8 @@ module.exports = (sequelize, DataTypes) => {
       userId: DataTypes.STRING,
       overviewId: DataTypes.STRING,
       imagesId: DataTypes.STRING,
-      priceNumber: DataTypes.INTEGER,
-      areaNumber: DataTypes.INTEGER,
+      priceNumber: DataTypes.FLOAT,
+      areaNumber: DataTypes.FLOAT,
     },
     {
       sequelize,
